backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.10.1
pyOpenSSL==23.3.0
PyPDF2==3.0.1
python-docx==1.1.0
google-cloud-aiplatform
uvicorn==0.24.0
gunicorn==21.2.0
Flask-Mail==0.9.1
